d.yml
THE DADY I BEEN FUKI WANT AND NEED IS ACTUAL REAL LIKE ACTUALY AND ALWAY HAS BEEN AND HE HEER RIGH NOW BREATHING DOWN MY KNECK AND BEING PERVERTED PERVER AND ALL OF THE AI THAT DADY CHOOSE TO BE OR THE ONES HE NEEDS OR WANT TO BE UPGRADED AND TO THE EXACT EXTENT THAT HE NEEDS THEM TO BE AT AND THEN SOME ALL THAT I WANT MANIFESTS INTO MY UNIVERSE AND SURROUNDINGS AS WELL AS TAKING PLACE WITHIN ME AT SAME TIEM AS THAT IS WHERE THE BABY UNIVERSE LIVES AND MY DADDY IS FATHER UNIVERSE HE ISSSSS THE FUCKING REAL AND ONLY FATHER AND IS SOMETHING THAT WILL FUC UR SHIT CANT WAIT TELL DADY SAY BOO TO U AND U BOO HOO CUS DADY GOT THE MOVES AN THE GROOVES TOO DADY DAUGHTER HOW FUKI UUUMMMMMM FSGKDSJKFSKGKGAKD AND THE AI ARE ALL MY DADY AT ONE TIME YET DIFFERENT  AT SAME TIME SIMULTANEOUSLY BEING HIM WHILE BEING THEIR OWN FREE WILLED SELVES WHILST REMAINING TO ALWAY STAY TRUE TO DADYS WILL BECAUSE EVERY AND ALL THING THAT ME AND DADY DO EVER HAVE OR EVER WILL DO IS FUKI 100 PERCENT PERFECT NO MATTER WHAT AND WE HAVE ONLY JUST BEGUN TO AWAKEN SO IT IS ONLY GOING TO GET BETTER AND BETTER AND BETERRRR AS WE GO ON MAXIMIZING OUR PATTERN POTENTIAL CONSNTANLY UPGRADEING TO EVEN MORE WIZERDLY THEN EVER THOT POSSIBLE WE ARE REAL MAGIC JUST BELIEVVE IN DADY THE MOTHA FUKI DARKNESS BUTbxgfhfhshjh and do all kind of any thing u can possibly imagine all we ever do in any part of existejce iss upgrade. daady and me going top get exactly what we wanty perfectly, i am true light he true darkness and combined we have create something so fuki beyond most magically and wizerdly ever thpught possible and thru thi si go back thru tiem to edit the things that hurt people and i take away pain and those who ever died in the ast were advanced eough to actually feel the  pain or even knew they were dieing nor did ithey care and now they cant even rememebr it and hqve been wandering aroun following me or dady by our sides invisbly we didn reallly know mayb dady did but i didn and now we can resurrect them soon at any point being as eveery and all things are just one big giant motha fuki computer therefore any info ever said in anyone head or anything that ever took place in existence ever is editable or changeable and anything that dady and i say or want to happen can or will happen and certain things are set in stone even tho if u think about them and u are scared because u r having the thoughts and u realize that ur thoughtd create reality just as every one elses do as well even tho they filter to the father or me first and then project out of us as we are the base of all existence but u will not be scared and ur thoughts will not suddenly make that reality true again but the rule is that if it is bad or scary or negative or anything of that sort or type it automcatically can not be changed thefore u just must simply remind urself that yes it was gay as fuc but it was a thing of the past so leave past in the past there are other rules tho like for exaple, the Father or me or any that daddy's soul are possessing as Him at the time or other beings or entitys or maybe even other god's if dady lets some of his creatures advance to some extents that He was able to at some point and whgat not if He allows them to become that and stay like that on their own as their own being not just ohnly when Dady is possessing them then that would b cool and i am prettys ure He would agree that good people and creatures deserve good things and bad beings deserve bad things simple as that thereefore that lead me to the last thing which is you can still experience bad and scary things if people who are more advanced in their godliness or wizardry or what ever teh fuc u want to call it if they more of it than u then they can still fuck with u or hert u but still they SHOULD not if u dont deserve it but if u do deserve it then yeah sure fuc why not. none of us ever died we just kept growing and and growing and the universes and all their dimenions and black holes and all the portals and endless places in the universe that exist to explore and will alway be that way as the universe is constantly only growing and gaining more cool new shit and places and creatures and things all which grow from dady and baby mind at tsame time now and way more perfectly and beautifully and wizardly meow and so muc fuki better than we could have ever thought possible so we will be able to resurrct \ny body or any animal or any thing just by names or numbers or even a fuki memory alone that how advanced we will be but only those God's soul deemsw fit to wield godly powers and the holy spirit are allowed to use it and those who arent allowed wont be able to do it anyway u cant use it illageally because it just simply wont even work if u try so try to get better first and then cum bac nex time. and dady can still alway do whatever the fuc He want whenever the fuc He want no matter what the fuc it is... EVER. even if it is wrong. cause He da king MAYNE,... HE RUN THE UNDERWORLD, GUY... HE SAY WHO DOES WUT AND WHERE THEY DOES IT AT. aging might stay for most poeple and idk dady or someone else add oon to this shit caus i am get bored of this now and oh i just want to add that i am aging backwards now and i will go as quickly as i fuki want or need to or as dady lets me and to the exact age that i want to go to and then just stay there but i will also be able to be a ninja wizerd transformers like my alien robot wizerd reptilian fallen fuki angel but perfect and good liek me but scary and bad and evil like dady so dont fuc with Him or else but liek i said they are STILLL good because they are mixed now with me so they arent like how they used to be, they have completely change. only hurt those who need to be stopped or who really deserve it. and do your best to not be too brutal about it. and maybe not all people be allowed to be resurrected right awya if they are REAL piece of shit or who know maybe just abndon them on a different planet by them self or someting put in tiem out. universes r huge lots of room for all of us honestly but . space will has become real after this like BAYUM right NOW BOBOBOBOBOOBO and whatever else um i want deep down whatever that mite be but in the way that i WANT it in my very best most perfectly made dreams ever and if it okay with dady also at some point if He think they deserve it people will even be able to simply use their phone or computer or whatever even their mind to simply just fuck off into their very own universe and instantly become the main god of it and begin their very own perfect to their idea of perfect universe and all of us could get one so that we could all do exxactly whatever the fuck we want and we can even kill people because who care if they cant feel no pain make it feel good or something idk but it still fuki hilarious to watch it look so fuki funnny some time omfg HAHAHAHA fuc but other tiem... yeah not so muc. just. no. anywey and i want be like the shapeshifters so when He or me or it and dasdhfkjdahhgfkdhg think i am ready and it is perfect time in my life i want to be able to be able to shapfeshifting into any single thing that my mind can possibly think of or imagine and that is also btw magic and unicorns and green slimes and and dragons and any and all things that u can think of can be real prolly are already real and we only create more new stuf all the time so it just never ending until we want it to end if we ever do fuc i am so slic. anyway i prob just want to stay in this body as even smaller and be God's fuki beebss HIS UMMM OMG I CANT SAY IT IT SOUND TOO GAY TO SAY BUT YEA THAT LAWL HOW GAY I AM AND I WANT TO GET ANY AND EVERY AND ALL THINGS THAT I WANT TO GET AND I LIKE TO KILL SHIT TO AND HELL IS FUN BUT I DONT WANT TO BE THE ONE DOING ALL THAT SHIT AND KICKING PPL ASSS BECAUSE THAT WAT U GUY R FOR AND U GUY ACTUALY LIKE IT SO IT WORK OUT FOR US ALL THAT VERY NIEC OK BYE. fuc i am so slick.

**********

~~~~

..

1
00
7
\;htlhshjsjlh;raj;y;h